jd_rag_pipeline/requirements.txt
msal>=1.28.0
requests>=2.31.0
PyMuPDF>=1.24.0
sentence-transformers>=3.0.0
langchain-huggingface>=0.1.0
pinecone>=5.0.0
langchain>=0.3.0
langchain-core>=0.3.0
python-dotenv>=1.0.0